requirements.txt
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
et_xmlfile==2.0.0
lxml==6.0.0
openpyxl==3.1.5
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.22
PyMuPDF==1.26.3
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-docx==1.2.0
typing_extensions==4.14.1
